@@ -4,21 +4,22 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   cache:
     path: ../cache
   equatable: ^2.0.3
-  firebase_auth: ^3.0.2
-  firebase_core: ^1.5.0
+  firebase_auth: ^3.6.2
+  firebase_core: ^1.20.0
   flutter:
     sdk: flutter
-  google_sign_in: ^5.0.7
-  meta: ^1.3.0
-  very_good_analysis: ^2.4.0
+  google_sign_in: ^5.4.0
+  meta: ^1.7.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mocktail: ^0.2.0
+  mocktail: ^0.3.0
+  plugin_platform_interface: ^2.1.2
+  very_good_analysis: ^3.0.1
