@@ -4,8 +4,8 @@ version: 1.0.0
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dev_dependencies:
   test: ^1.17.11
-  very_good_analysis: ^2.4.0
+  very_good_analysis: ^3.0.1
