@@ -4,7 +4,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   authentication_repository:
@@ -12,24 +12,24 @@ dependencies:
   bloc:
     path: ../../packages/bloc
   equatable: ^2.0.3
-  flow_builder: ^0.0.4
+  flow_builder: ^0.0.9
   flutter:
     sdk: flutter
   flutter_bloc:
     path: ../../packages/flutter_bloc
-  font_awesome_flutter: ^9.1.0
+  font_awesome_flutter: ^10.1.0
   form_inputs:
     path: packages/form_inputs
   formz: ^0.4.0
-  google_fonts: ^2.1.0
-  meta: ^1.3.0
-  very_good_analysis: ^2.4.0
+  google_fonts: ^3.0.1
+  meta: ^1.7.0
 
 dev_dependencies:
   bloc_test: ^9.0.0
   flutter_test:
     sdk: flutter
-  mocktail: ^0.2.0
+  mocktail: ^0.3.0
+  very_good_analysis: ^3.0.1
 
 flutter:
   uses-material-design: true
