@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 const _avatarSize = 48.0;
 
 class Avatar extends StatelessWidget {
-  const Avatar({Key? key, this.photo}) : super(key: key);
+  const Avatar({super.key, this.photo});
 
   final String? photo;
 
