@@ -3,7 +3,10 @@ import 'package:flutter_firebase_login/app/app.dart';
 import 'package:flutter_firebase_login/home/home.dart';
 import 'package:flutter_firebase_login/login/login.dart';
 
-List<Page> onGenerateAppViewPages(AppStatus state, List<Page<dynamic>> pages) {
+List<Page<dynamic>> onGenerateAppViewPages(
+  AppStatus state,
+  List<Page<dynamic>> pages,
+) {
   switch (state) {
     case AppStatus.authenticated:
       return [HomePage.page()];
