@@ -14,7 +14,7 @@ class Email extends FormzInput<String, EmailValidationError> {
   const Email.pure() : super.pure('');
 
   /// {@macro email}
-  const Email.dirty([String value = '']) : super.dirty(value);
+  const Email.dirty([super.value = '']) : super.dirty();
 
   static final RegExp _emailRegExp = RegExp(
     r'^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$',
