@@ -14,7 +14,7 @@ class Password extends FormzInput<String, PasswordValidationError> {
   const Password.pure() : super.pure('');
 
   /// {@macro password}
-  const Password.dirty([String value = '']) : super.dirty(value);
+  const Password.dirty([super.value = '']) : super.dirty();
 
   static final _passwordRegExp =
       RegExp(r'^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$');
