@@ -9,14 +9,26 @@ void main() {
     test('returns [HomePage] when authenticated', () {
       expect(
         onGenerateAppViewPages(AppStatus.authenticated, []),
-        [isA<MaterialPage>().having((p) => p.child, 'child', isA<HomePage>())],
+        [
+          isA<MaterialPage<void>>().having(
+            (p) => p.child,
+            'child',
+            isA<HomePage>(),
+          )
+        ],
       );
     });
 
     test('returns [LoginPage] when unauthenticated', () {
       expect(
         onGenerateAppViewPages(AppStatus.unauthenticated, []),
-        [isA<MaterialPage>().having((p) => p.child, 'child', isA<LoginPage>())],
+        [
+          isA<MaterialPage<void>>().having(
+            (p) => p.child,
+            'child',
+            isA<LoginPage>(),
+          )
+        ],
       );
     });
   });
