@@ -11,7 +11,7 @@ class MockAuthenticationRepository extends Mock
 void main() {
   group('LoginPage', () {
     test('has a page', () {
-      expect(LoginPage.page(), isA<MaterialPage>());
+      expect(LoginPage.page(), isA<MaterialPage<void>>());
     });
 
     testWidgets('renders a LoginForm', (tester) async {
